requirements.txt
torch --index-url https://download.pytorch.org/whl/cu128
torchtext
sacrebleu
tqdm
sentencepiece
